third_party/WebKit/LayoutTests/accessibility/aom-computed-accessible-node.html

<!DOCTYPE HTML>
<script src="../resources/gc.js"></script>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<!--

Accessibility Object Model
Explainer: https://github.com/WICG/aom/blob/master/explainer.md
Spec: https://wicg.github.io/aom/spec/

-->

<div id="button1" role="button">Click</div>

<script>

test(function(t) {
    assert_true(internals.runtimeFlags.accessibilityObjectModelEnabled);
}, "Make sure that Accessibility Object Model is enabled");

test(async function(test) {
    var element = document.getElementById("button1");
    var computedAXNode = await window.getComputedAccessibleNode(element)
    assert_true(computedAXNode != null);
}, 'Ensure that a non null value is returned from getComputedAccessibleNode');


test(async function(test) {
    var element = document.getElementById("button1");

    var computedAXNode1 = await window.getComputedAccessibleNode(element);
    var computedAXNode2 = await window.getComputedAccessibleNode(element);
    assert_true(computedAXNode1 === computedAXNode2);
}, "Multiple calls to getComputedAccessibleNode will return the same node.");
</script>

<div id="button2" role="button" aria-label="axButton">Click</div>

<script>

test(async function(test) {
    var button1 = document.getElementById("button1");
    var button1CAXNode = await window.getComputedAccessibleNode(button1);

    assert_equals(button1CAXNode.name, "Click");
    assert_equals(button1CAXNode.role, "button");

    button1.style.display = "none";

    var button2 = document.getElementById("button2");
    var button2CAXNode = await window.getComputedAccessibleNode(button2);

    assert_equals(button2CAXNode.name, "axButton");
    assert_equals(button2CAXNode.role, "button");

    // As button1 has no node in the accessibility tree anymore, assert that the
    // its previously retrieved computed accessible node has had its attributes
    // nullified.
    assert_equals(button1CAXNode.name, null);
    assert_equals(button1CAXNode.role, null);

}, "Deleting nodes from the accessibility tree will not cause a crash, and properties on any references to a deleted computed accessible node have been nullified.");

</script>

<div id="radio1" role="radio">Radio 1</div>
<div id="radio2" role="radio">Radio 2</div>

<script>

async_test(async function(test) {
    var radio1 = document.getElementById("radio1");
    var caxNode1 = await window.getComputedAccessibleNode(radio1);
    var radio2 = document.getElementById("radio2");
    var caxNode2 = await window.getComputedAccessibleNode(radio2);
    radio1.innerHTML = "pizza";
    radio2.innerHTML = "tacos";

    // ComputedAccessibleNode for this element has not been updated so should
    // only reflect the name at the time of its request.
    assert_equals(caxNode1.name, "Radio 1");
    assert_equals(caxNode2.name, "Radio 2");

    // Manually refresh the node and check that all references to
    // ComputedAccessibleNodes have been updated.
    await caxNode1.ensureUpToDate();
    assert_equals(caxNode1.name, "pizza");
    assert_equals(caxNode2.name, "tacos");

    radio1.setAttribute("aria-label", "pasta");
    assert_equals(caxNode1.name, "pizza");

    caxNode1.ensureUpToDate();
    assert_equals(caxNode1.name, "pasta");

    test.done();
}, "ComputedAccessibleNode.ensureUpToDate");



</script>
